refactor: move render_template to roz-core
Shared by roz-local (REST daemon tools) and roz-copper (WS bridge).
roz-local re-exports for backward compatibility.
